refactor(modpath): refactor particledata
Refactor ParticleData (input style 1) so that it is the same as the node template (input style 3).
