+++ README.md
@@ -1,10 +1,10 @@
 
 <img src="https://raw.githubusercontent.com/modflowpy/flopy/master/examples/images/flopy3.png" alt="flopy3" style="width:50;height:20">
 
-### Version 3.2.9 mp7-refactor &mdash; build 256
-[![Build Status](https://travis-ci.org/modflowpy/flopy.svg?branch=mp7-refactor)](https://travis-ci.org/modflowpy/flopy)
+### Version 3.2.9 mp7part-refactor &mdash; build 213
+[![Build Status](https://travis-ci.org/modflowpy/flopy.svg?branch=mp7part-refactor)](https://travis-ci.org/modflowpy/flopy)
 [![PyPI Version](https://img.shields.io/pypi/v/flopy.png)](https://pypi.python.org/pypi/flopy)
-[![Coverage Status](https://coveralls.io/repos/github/modflowpy/flopy/badge.svg?branch=mp7-refactor)](https://coveralls.io/github/modflowpy/flopy?branch=mp7-refactor)
+[![Coverage Status](https://coveralls.io/repos/github/modflowpy/flopy/badge.svg?branch=mp7part-refactor)](https://coveralls.io/github/modflowpy/flopy?branch=mp7part-refactor)
 
 
 Introduction
@@ -174,7 +174,7 @@ How to Cite
 
 ##### ***Software/Code citation for FloPy:***
 
-[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7-refactor: U.S. Geological Survey Software Release, 12 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
+[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7part-refactor: U.S. Geological Survey Software Release, 13 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
 
 
 MODFLOW Resources
