+++ code.json
@@ -29,9 +29,9 @@
         "downloadURL": "https://code.usgs.gov/usgs/modflow/flopy/archive/master.zip", 
         "vcs": "git", 
         "laborHours": -1, 
-        "version": "3.2.9.256", 
+        "version": "3.2.9.213", 
         "date": {
-            "metadataLastUpdated": "2018-10-12"
+            "metadataLastUpdated": "2018-10-13"
         }, 
         "organization": "U.S. Geological Survey", 
         "permissions": {
