+++ docs/PyPi_release.md
@@ -29,7 +29,7 @@ How to Cite
 
 *Software/Code citation for FloPy:*
 
-[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7-refactor: U.S. Geological Survey Software Release, 12 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
+[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7part-refactor: U.S. Geological Survey Software Release, 13 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
 
 
 Disclaimer
@@ -46,12 +46,12 @@ resulting from the authorized or unauthorized use of the software.
 
 Installation
 -----------------------------------------------
-To install FloPy version 3.2.9 &mdash; mp7-refactor:
+To install FloPy version 3.2.9 &mdash; mp7part-refactor:
 ```
 pip install flopy
 ```
 
-To update to FloPy version 3.2.9 &mdash; mp7-refactor:
+To update to FloPy version 3.2.9 &mdash; mp7part-refactor:
 ```
 pip install flopy --upgrade
 ```
