+++ docs/USGS_release.md
@@ -17,9 +17,9 @@ header-includes:
     - \fancyhf{{}}
     - \fancyhead[LE, LO, RE, RO]{}
     - \fancyhead[CE, CO]{FloPy Release Notes}
-    - \fancyfoot[LE, RO]{FloPy version 3.2.9 &mdash; mp7-refactor}
+    - \fancyfoot[LE, RO]{FloPy version 3.2.9 &mdash; mp7part-refactor}
     - \fancyfoot[CO, CE]{\thepage\ of \pageref{LastPage}}
-    - \fancyfoot[RE, LO]{10/12/2018}
+    - \fancyfoot[RE, LO]{10/13/2018}
 geometry: margin=0.75in
 ---
 
@@ -54,7 +54,7 @@ How to Cite
 
 ##### ***Software/Code citation for FloPy:***
 
-[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7-refactor: U.S. Geological Survey Software Release, 12 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
+[Bakker, M., Post, V., Langevin, C.D., Hughes, J.D., White, J.T., Starn, J.J., and Fienen, M.N., 2018, FloPy v3.2.9 &mdash; mp7part-refactor: U.S. Geological Survey Software Release, 13 October 2018, http://dx.doi.org/10.5066/F7BK19FH](http://dx.doi.org/10.5066/F7BK19FH)
 
 
 Disclaimer
@@ -71,12 +71,12 @@ resulting from the authorized or unauthorized use of the software.
 
 Installation
 -----------------------------------------------
-To install FloPy version 3.2.9 &mdash; mp7-refactor from the USGS FloPy website:
+To install FloPy version 3.2.9 &mdash; mp7part-refactor from the USGS FloPy website:
 ```
 pip install https://water.usgs.gov/ogw/flopy/flopy-3.2.9.zip
 ```
 
-To update to FloPy version 3.2.9 &mdash; mp7-refactor from the USGS FloPy website:
+To update to FloPy version 3.2.9 &mdash; mp7part-refactor from the USGS FloPy website:
 ```
 pip install https://water.usgs.gov/ogw/flopy/flopy-3.2.9.zip --upgrade
 ```
