+++ flopy/modpath/__init__.py
@@ -4,5 +4,7 @@ from .mpsim import ModpathSim
 from .mp7 import Modpath7
 from .mp7bas import Modpath7Bas
 from .mp7sim import Modpath7Sim
-from .mp7particle import ParticleGroup, ParticleGroupNodeTemplate, \
-    NodeParticleDataFace, NodeParticleDataCell
+from .mp7particlegroup import ParticleGroup, ParticleGroupLRCTemplate, \
+    ParticleGroupNodeTemplate
+from .mp7particledata import ParticleData, NodeParticleDataFace, \
+    NodeParticleDataCell
