+++ flopy/version.py
@@ -1,11 +1,11 @@
 # flopy version file automatically created using...pre-commit.py
-# created on...October 12, 2018 13:51:57
+# created on...October 13, 2018 15:47:59
 
 major = 3
 minor = 2
 micro = 9
-build = 256
-commit = 2927
+build = 213
+commit = 2884
 
 __version__ = '{:d}.{:d}.{:d}'.format(major, minor, micro)
 __build__ = '{:d}.{:d}.{:d}.{:d}'.format(major, minor, micro, build)
